[OPR-99]: Makes things compile for OS X
