+++ Operations/Extras/AddressBook/Shared/Contacts.swift
@@ -221,8 +221,36 @@ public struct SystemContactStore: ContactStoreType {
     }
 }
 
+@available(iOS 9.0, OSX 10.11, *)
+extension SystemContactStore: AddressBookRegistrarType {
 
+    public func opr_authorizationStatusForRequirement(entityType: AddressBookAuthorizationStatus.EntityType) -> AddressBookAuthorizationStatus {
+        return opr_authorizationStatusForEntityType(CNEntityType(entity: entityType)).addressBookAuthorizationStatus
+    }
 
+    public func opr_requestAccessForRequirement(entityType: AddressBookAuthorizationStatus.EntityType, completion: (Bool, NSError?) -> Void) {
+        opr_requestAccessForEntityType(CNEntityType(entity: entityType), completion: completion)
+    }
+}
+
+@available(iOS 9.0, OSX 10.11, *)
+extension CNAuthorizationStatus {
+    var addressBookAuthorizationStatus: AddressBookAuthorizationStatus {
+        switch self {
+        case .NotDetermined: return .NotDetermined
+        case .Restricted: return .Restricted
+        case .Denied: return .Denied
+        case .Authorized: return .Authorized
+        }
+    }
+}
+
+@available(iOS 9.0, OSX 10.11, *)
+extension CNEntityType {
+    init(entity: AddressBookAuthorizationStatus.EntityType) {
+        self = .Contacts
+    }
+}
 
 @available(iOS 9.0, OSX 10.11, *)
 extension ContainerID: CustomStringConvertible {
