+++ OperationsTests/Extras/AddressBook/AddressBookTests.swift
@@ -297,6 +297,27 @@ class AddressBooksTests: XCTestCase {
     }
 }
 
+@available(iOS, deprecated=9.0)
+class ABAuthorizationStatusTests: XCTestCase {
+
+    func test__given_authorized__address_book_authorization_status_is_correct() {
+        XCTAssertEqual(ABAuthorizationStatus.Authorized.addressBookAuthorizationStatus, AddressBookAuthorizationStatus.Authorized)
+    }
+
+    func test__given_not_determined__address_book_authorization_status_is_correct() {
+        XCTAssertEqual(ABAuthorizationStatus.NotDetermined.addressBookAuthorizationStatus, AddressBookAuthorizationStatus.NotDetermined)
+    }
+
+    func test__given_denied__address_book_authorization_status_is_correct() {
+        XCTAssertEqual(ABAuthorizationStatus.Denied.addressBookAuthorizationStatus, AddressBookAuthorizationStatus.Denied)
+    }
+
+    func test__given_restricted__address_book_authorization_status_is_correct() {
+        XCTAssertEqual(ABAuthorizationStatus.Restricted.addressBookAuthorizationStatus, AddressBookAuthorizationStatus.Restricted)
+    }
+}
+
+
